requirements.txt
streamlit
supabase
python-dotenv
pyfunda
pypararius
resend
datetime